requirements.txt
Django==5.0.7
pillow==10.4.0
pytesseract==0.3.10
pdf2image==1.17.0
